requirements.txt
fastapi
sqlalchemy
uvicorn
pydantic
pydantic[email]
pydantic_settings
psycopg2-binary
alembic
passlib
python-jose
python-multipart
httpx
bcrypt==3.2.2
google-api-python-client
google-genai
qdrant_client